fix 🐛: Update dependencies and add HTTPx to project requirements
This commit updates several dependencies in the project, including `httpx` from version 0.28.1 to a more recent version (0.29.0). Additionally, it adds `httpx` to the list of required packages in `pyproject.toml`, `requirements.dev.txt`, and `requirements.txt`. These changes ensure that the project remains up-to-date with the latest HTTP client features and improvements, enhancing functionality and compatibility.